+++ features/step_definitions/dump_flows_steps.rb
@@ -1,9 +1,8 @@
 Then(/^the switch "(.*?)" has (\d+) flow entr(?:y|ies)$/) do |switch, number|
-  # FIXME: Read TREMA_SOCKET_DIR in trema dump_flows
-  command = "trema dump_flows #{switch} -S ."
+  command = "trema dump_flows #{switch}"
   step "I successfully run `#{command}`"
   dump_flows = aruba.command_monitor.find(Aruba.platform.detect_ruby(command))
-  expect(dump_flows.output.split("\n").size - 1).to eq(number.to_i)
+  expect(dump_flows.output.split("\n").size).to eq(number.to_i)
 end
 
 Then(/^the switch "(.*?)" has no flow entry$/) do |switch|