+++ lib/trema/drb.rb
@@ -5,18 +5,24 @@ require 'phut'
 module Trema
   def self.trema_process(controller_name, socket_dir)
     Phut.socket_dir = socket_dir
-    socket_path = File.join(Phut.socket_dir, "trema.#{controller_name}.ctl")
+    socket_path = File.join(Phut.socket_dir, "#{controller_name}.ctl")
     unless FileTest.socket?(socket_path)
-      fail %(Controller process "#{controller_name}" does not exist.)
+      raise %(Controller process "#{controller_name}" does not exist.)
     end
     DRbObject.new_with_uri('drbunix:' + socket_path)
   end
 
   def self.trema_processes(socket_dir = Phut.socket_dir)
     Phut.socket_dir = socket_dir
-    Dir.glob(File.join Phut.socket_dir, 'trema.*.ctl').map do |each|
-      DRbObject.new_with_uri('drbunix:' + each)
-    end
+    all = Dir.glob(File.join(Phut.socket_dir, '*.ctl'))
+    vhosts = Dir.glob(File.join(Phut.socket_dir, 'vhost.*.ctl'))
+    (all - vhosts).map { |each| DRbObject.new_with_uri("drbunix:#{each}") }
+  end
+
+  def self.vhosts(socket_dir = Phut.socket_dir)
+    Phut.socket_dir = socket_dir
+    vhosts = Dir.glob(File.join(Phut.socket_dir, 'vhost.*.ctl'))
+    vhosts.map { |each| DRbObject.new_with_uri("drbunix:#{each}") }
   end
 
   def self.fetch(name, socket_dir)
@@ -27,6 +33,6 @@ module Trema
         next
       end
     end
-    fail %("#{name}" does not exist.)
+    raise %("#{name}" does not exist.)
   end
 end