+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## develop (unreleased)
 
 
+## 0.10.0 (2/17/2015)
+### New features
+* [#418](https://github.com/trema/trema/pull/418): Ruby 2.3.0 support
+* [#419](https://github.com/trema/trema/pull/419): Add trema reset_stats command
+* [#421](https://github.com/trema/trema/pull/421): Add Trema::Logger#unknown method
+
+
 ## 0.9.0 (12/17/2015)
 ### New features
 * [#415](https://github.com/trema/trema/pull/415): Add `trema netns [command]` command.